Several improvements
Fixed bug breaking explicit adams
Fixed wrong atol in odeint
Added dtype casting after interacting with tableaus etc.
Added correct dtype to tableau etc. tensors
Added version number to source.

+++ torchdiffeq/_impl/adaptive_heun.py
@@ -4,20 +4,20 @@ from .solvers import RKAdaptiveStepsizeODESolver
 
 
 _ADAPTIVE_HEUN_TABLEAU = _ButcherTableau(
-    alpha=torch.tensor([1.]),
+    alpha=torch.tensor([1.], dtype=torch.float64),
     beta=[
-        torch.tensor([1.]),
+        torch.tensor([1.], dtype=torch.float64),
     ],
-    c_sol=torch.tensor([0.5, 0.5]),
+    c_sol=torch.tensor([0.5, 0.5], dtype=torch.float64),
     c_error=torch.tensor([
         0.5,
         -0.5,
-    ]),
+    ], dtype=torch.float64),
 )
 
 _AH_C_MID = torch.tensor([
     0.5, 0.
-])
+], dtype=torch.float64)
 
 
 class AdaptiveHeunSolver(RKAdaptiveStepsizeODESolver):
